+++ src/main/java/com/atlauncher/utils/walker/CaseFileVisitor.java
@@ -23,7 +23,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.attribute.BasicFileAttributes;
-import java.util.ArrayList;
 import java.util.List;
 
 import com.atlauncher.data.json.CaseType;